1规范部分代码 将原有继承Fragment变为继承BaseFragment 2 修改BaseActivity部分代码(增加handler变量，修复跨进程获取服务异步空指针问题) 3 在BaseFragment中增加getPermission方法(获取用户当前权限 返回值为boolean)

+++ app/src/main/java/cn/ml_tech/mx/mlproj/YpkFragment.java
@@ -1,17 +1,16 @@
 package cn.ml_tech.mx.mlproj;
 
-import android.app.Fragment;
-import android.content.Context;
-import android.net.Uri;
 import android.os.Bundle;
 import android.os.RemoteException;
+import android.support.annotation.Nullable;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
-import android.view.ViewGroup;
+import android.widget.Button;
 import android.widget.EditText;
+import android.widget.ImageButton;
+import android.widget.LinearLayout;
 import android.widget.TextView;
 import android.widget.Toast;
 
@@ -24,29 +23,29 @@ import cn.ml_tech.mx.mlservice.IMlService;
 
 import static cn.ml_tech.mx.mlproj.R.id.me_name;
 
+public class YpkFragment extends BaseFragment{
 
-/**
- * A simple {@link Fragment} subclass.
- * Activities that contain this fragment must implement the
- * {@link YpkFragment.OnFragmentInteractionListener} interface
- * to handle interaction events.
- * Use the {@link YpkFragment#newInstance} factory method to
- * create an instance of this fragment.
- */
-public class YpkFragment extends Fragment {
-    // TODO: Rename parameter arguments, choose names that match
-    // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-    private static final String ARG_PARAM1 = "param1";
-    private static final String ARG_PARAM2 = "param2";
-
-    // TODO: Rename and change types of parameters
-    private String mParam1;
-    private String mParam2;
 
     private List<DrugControls> drugList = new ArrayList<>();
     private DrugControls drugControls = null;
     private DrugAdapter adapter;
     private YpjcActivity ypjcActivity;
+    private View view;
+    private EditText meName;
+    private Button query;
+    private EditText mePhonetic;
+    private Button reseting;
+    private EditText meEnname;
+    private Button addphonetic;
+    private LinearLayout wholelayout;
+    private ImageButton ibPre;
+    private EditText etPage;
+    private ImageButton ibNext;
+    private TextView tvCurrentPage;
+    private TextView tvAllPage;
+    private ImageButton ibSearch;
+    private Button btnypxNext;
+
 
     public IMlService getmService() {
         return mService;
@@ -61,27 +60,28 @@ public class YpkFragment extends Fragment {
     RecyclerView recyclerView;
     private IMlService mService;
     private String mDrugName = "", mPinyin = "", mEnName = "";
-    private OnFragmentInteractionListener mListener;
     private DrugAdapter.OperateToData operateToData = new DrugAdapter.OperateToData() {
         @Override
         public boolean delete(long id) {
-            Log.d("zw", id + "id");
-            try {
-                getmService().deleteDrugInfoById((int) id);
-                getmService().deleteDrugParamById((int) id);
-                adapter.deleteDataById((int) id);
-                Toast.makeText(getActivity(), "删除成功", Toast.LENGTH_SHORT).show();
-            } catch (RemoteException e) {
-                e.printStackTrace();
-                Toast.makeText(getActivity(), "删除成功 原因:" + e.getLocalizedMessage(), Toast.LENGTH_SHORT).show();
-
+            if (getPermissionById(9, 5)) {
+                try {
+                    getmService().deleteDrugInfoById((int) id);
+                    getmService().deleteDrugParamById((int) id);
+                    adapter.deleteDataById((int) id);
+                    Toast.makeText(getActivity(), "删除成功", Toast.LENGTH_SHORT).show();
+                } catch (RemoteException e) {
+                    e.printStackTrace();
+                    Toast.makeText(getActivity(), "删除成功 原因:" + e.getLocalizedMessage(), Toast.LENGTH_SHORT).show();
+                }
+            } else {
+                showRefuseTip();
             }
             return false;
         }
 
         @Override
         public void operateToPre(boolean isNext, DrugControls drugControl) {
-            getActivity().findViewById(R.id.btnypxNext).setEnabled(isNext);
+            btnypxNext.setEnabled(isNext);
             ypjcActivity.drugControl = drugControl;
             ypjcActivity.druginfo_id = (int) drugControl.getId();
 
@@ -89,74 +89,89 @@ public class YpkFragment extends Fragment {
 
         @Override
         public void update(DrugControls drugControls) {
-            ypjcActivity.druginfo_id = (int) drugControls.getId();
-            ypjcActivity.drugControl = drugControls;
-            ypjcActivity.moveToAddDrug();
+            if (getPermissionById(9, 4)) {
+                ypjcActivity.druginfo_id = (int) drugControls.getId();
+                ypjcActivity.drugControl = drugControls;
+                ypjcActivity.moveToAddDrug();
+            } else {
+                showRefuseTip();
+            }
         }
     };
-
-    public YpkFragment() {
-        // Required empty public constructor
-    }
-
-    /**
-     * Use this factory method to create a new instance of
-     * this fragment using the provided parameters.
-     *
-     * @param param1 Parameter 1.
-     * @param param2 Parameter 2.
-     * @return A new instance of fragment YpkFragment.
-     */
-    // TODO: Rename and change types and number of parameters
-    public static YpkFragment newInstance(String param1, String param2) {
-        YpkFragment fragment = new YpkFragment();
-        Bundle args = new Bundle();
-        args.putString(ARG_PARAM1, param1);
-        args.putString(ARG_PARAM2, param2);
-        fragment.setArguments(args);
-        return fragment;
-    }
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        if (getArguments() != null) {
-            mParam1 = getArguments().getString(ARG_PARAM1);
-            mParam2 = getArguments().getString(ARG_PARAM2);
-        }
-    }
 
     private void initDrugs() {
         try {
             drugList = mService.queryDrugControl();
+            setDataToView(drugList, true);
         } catch (RemoteException e) {
             e.printStackTrace();
         }
+
     }
 
-    // etPage 页数输入框
-    // ibPre imagebutton 上一页
-    //ibNext imageButton 下一页
-    //tvCurrentPage 当前页数
-    //tvAllPage 总页数
-    // ibSearch 查询
     @Override
     public void onStart() {
         super.onStart();
         ypjcActivity = (YpjcActivity) getActivity();
-        initDrugs();
-        getActivity().findViewById(R.id.ibPre).setOnClickListener((View.OnClickListener) getActivity());
-        getActivity().findViewById(R.id.ibNext).setOnClickListener((View.OnClickListener) getActivity());
-        getActivity().findViewById(R.id.ibSearch).setOnClickListener((View.OnClickListener) getActivity());
-        getActivity().findViewById(R.id.btnypxNext).setEnabled(false);
-        getActivity().findViewById(R.id.query).setOnClickListener((View.OnClickListener) getActivity());
-        getActivity().findViewById(R.id.reseting).setOnClickListener((View.OnClickListener) getActivity());
-        getActivity().findViewById(R.id.btnypxNext).setOnClickListener((View.OnClickListener) getActivity());
-        getActivity().findViewById(R.id.addphonetic).setOnClickListener((View.OnClickListener) getActivity());
-        recyclerView = (RecyclerView) getActivity().findViewById(R.id.recycler_view);
+    }
+
+    @Override
+    public View initView(LayoutInflater inflater) {
+        view = inflater.inflate(R.layout.fragment_ypk, null);
+        initFindViewById(view);
+        return view;
+    }
+
+    @Override
+    public void initFindViewById(View view) {
+
+        meName = (EditText) view.findViewById(R.id.me_name);
+        query = (Button) view.findViewById(R.id.query);
+        mePhonetic = (EditText) view.findViewById(R.id.me_phonetic);
+        reseting = (Button) view.findViewById(R.id.reseting);
+        meEnname = (EditText) view.findViewById(R.id.me_enname);
+        addphonetic = (Button) view.findViewById(R.id.addphonetic);
+        wholelayout = (LinearLayout) view.findViewById(R.id.wholelayout);
+        recyclerView = (RecyclerView) view.findViewById(R.id.recycler_view);
+        ibPre = (ImageButton) view.findViewById(R.id.ibPre);
+        etPage = (EditText) view.findViewById(R.id.etPage);
+        ibNext = (ImageButton) view.findViewById(R.id.ibNext);
+        tvCurrentPage = (TextView) view.findViewById(R.id.tvCurrentPage);
+        tvAllPage = (TextView) view.findViewById(R.id.tvAllPage);
+        ibSearch = (ImageButton) view.findViewById(R.id.ibSearch);
+        btnypxNext = (Button) view.findViewById(R.id.btnypxNext);
+        recyclerView = (RecyclerView) view.findViewById(R.id.recycler_view);
         LinearLayoutManager linearLayoutManager = new LinearLayoutManager(getActivity());
         recyclerView.setLayoutManager(linearLayoutManager);
-        setDataToView(drugList, true);
+
+    }
+
+    @Override
+    protected void initEvent() {
+        super.initEvent();
+        view.findViewById(R.id.ibPre).setOnClickListener((View.OnClickListener) getActivity());
+        view.findViewById(R.id.ibNext).setOnClickListener((View.OnClickListener) getActivity());
+        view.findViewById(R.id.ibSearch).setOnClickListener((View.OnClickListener) getActivity());
+        view.findViewById(R.id.btnypxNext).setEnabled(false);
+        view.findViewById(R.id.query).setOnClickListener((View.OnClickListener) getActivity());
+        view.findViewById(R.id.reseting).setOnClickListener((View.OnClickListener) getActivity());
+        view.findViewById(R.id.btnypxNext).setOnClickListener((View.OnClickListener) getActivity());
+        view.findViewById(R.id.addphonetic).setOnClickListener((View.OnClickListener) getActivity());
+    }
+
+    /**
+     * @param savedInstanceState
+     */
+    @Override
+    public void initData(@Nullable Bundle savedInstanceState) {
+        setPermission(((YpjcActivity) getActivity()).permission);
+        initDrugs();
+        query.setVisibility(getPermissionById(7, 1) == true ? View.VISIBLE : View.INVISIBLE);
+        query.setEnabled(getPermissionById(7, 8));
+        addphonetic.setVisibility(getPermissionById(8, 1) == true ? View.VISIBLE : View.INVISIBLE);
+        addphonetic.setEnabled(getPermissionById(8, 8));
+        btnypxNext.setVisibility(getPermissionById(10, 1) == true ? View.VISIBLE : View.INVISIBLE);
+
     }
 
     public void setPreDataToView() throws RemoteException {
@@ -206,66 +221,28 @@ public class YpkFragment extends Fragment {
     }
 
     public void setDataToView(List<DrugControls> drugList, boolean isReseting) {
-        lastPage = ((int) Math.floor(drugList.size() / 20)) + 1;
-        ((TextView) getActivity().findViewById(R.id.tvAllPage)).setText(lastPage + "");
-        if (isReseting) {
-            cuurentPage = 1;
-            ((TextView) getActivity().findViewById(R.id.tvCurrentPage)).setText(cuurentPage + "/");
-            for (int i = 20; i < drugList.size(); i++) {
-                drugList.remove(i);
-                i--;
+        if (getPermissionById(9, 1)) {
+            lastPage = ((int) Math.floor(drugList.size() / 20)) + 1;
+            ((TextView) view.findViewById(R.id.tvAllPage)).setText(lastPage + "");
+            if (isReseting) {
+                cuurentPage = 1;
+                ((TextView) view.findViewById(R.id.tvCurrentPage)).setText(cuurentPage + "/");
+                for (int i = 20; i < drugList.size(); i++) {
+                    drugList.remove(i);
+                    i--;
+                }
             }
-        }
-
-        adapter = new DrugAdapter(drugList, getActivity(), operateToData);
-        recyclerView.setAdapter(adapter);
-    }
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_ypk, container, false);
-    }
-
-    // TODO: Rename method, update argument and hook method into UI event
-    public void onButtonPressed(Uri uri) {
-        if (mListener != null) {
-            mListener.onFragmentInteraction(uri);
+            adapter = new DrugAdapter(drugList, getActivity(), operateToData);
+            recyclerView.setAdapter(adapter);
         }
     }
 
-    @Override
-    public void onAttach(Context context) {
-        super.onAttach(context);
-        if (context instanceof OnFragmentInteractionListener) {
-            mListener = (OnFragmentInteractionListener) context;
-        } else {
-            throw new RuntimeException(context.toString()
-                    + " must implement OnFragmentInteractionListener");
-        }
-    }
 
     @Override
     public void onDetach() {
         super.onDetach();
-        mListener = null;
     }
 
-    /**
-     * This interface must be implemented by activities that contain this
-     * fragment to allow an interaction in this fragment to be communicated
-     * to the activity and potentially other fragments contained in that
-     * activity.
-     * <p>
-     * See the Android Training lesson <a href=
-     * "http://developer.android.com/training/basics/fragments/communicating.html"
-     * >Communicating with Other Fragments</a> for more information.
-     */
-    public interface OnFragmentInteractionListener {
-        // TODO: Update argument type and name
-        void onFragmentInteraction(Uri uri);
-    }
 
     public String getDrugName() {
         return ((EditText) getActivity().findViewById(me_name)).getEditableText().toString();
@@ -278,4 +255,6 @@ public class YpkFragment extends Fragment {
     public String getEnName() {
         return ((EditText) getActivity().findViewById(R.id.me_enname)).getEditableText().toString();
     }
+
+
 }
